+++ crates/ln-dlc-node/src/ln_dlc_wallet.rs
@@ -122,6 +122,10 @@ impl<S: TenTenOneStorage, N: Storage> LnDlcWallet<S, N> {
         self.address_cache.read().clone()
     }
 
+    pub fn update_lookahead(&self, gap: u32) -> Result<bool> {
+        self.ldk_wallet().update_lookahead(gap)
+    }
+
     pub fn new_address(&self) -> Result<Address> {
         self.ldk_wallet().get_new_address()
     }